fixes and new route

+++ src/main/java/br/com/github/aelkz/gmapseta/app/repository/Point.java
@@ -41,7 +41,7 @@ public enum Point {
 
     public static Point getPoint(Integer id) {
         for (Point point: Point.values()) {
-            if (point.getId().intValue() == id) {
+            if (point.getId().intValue() == id.intValue()) {
                 return point;
             }
         }
